+++ lib/src/scroll_notification_publisher.dart
@@ -4,11 +4,13 @@ import 'package:flutter/material.dart';
 
 // 共享滑动信息
 class ScrollNotificationPublisher extends InheritedWidget {
-  ScrollNotificationPublisher({Key? key, required Widget child})
+  ScrollNotificationPublisher(
+      {Key? key,
+      required Widget child,
+      required this.scrollNotificationController})
       : super(key: key, child: child);
 
-  final StreamController<ScrollNotification> scrollNotificationController =
-      StreamController<ScrollNotification>.broadcast();
+  final StreamController<ScrollNotification> scrollNotificationController;
 
   @override
   bool updateShouldNotify(covariant InheritedWidget oldWidget) {