+++ CHANGELOG.md
@@ -9,3 +9,7 @@
 * fix a null safety issue to adapt flutter 3.0.5
 * fix exposure factor doesn't work in first frame
 * support expose check manually
+
+### 0.0.4
+* adapt flutter 3
+* fix StreamController is change when setState called